[Refactor] Moved function and variable tracking to Exec.Lookup module"
