+++ GLaDOS2023.cabal
@@ -27,7 +27,8 @@ library
         Parsing.Token,
         Parsing.Ast,
         Parsing.Cpt,
-        Exec
+        Exec,
+        Exec.Lookup
     hs-source-dirs:   lib
     build-depends:    base >= 4.14.1.0, containers
     default-language: Haskell2010
@@ -55,7 +56,7 @@ test-suite test
     other-modules: Parsing.ArgsTests,
         Parsing.TokenTests,
         Parsing.CptTests,
-        Parsing.AstTests
+        Parsing.AstTests,
         ParsingTests,
         ExecTests
